add form component and implement refactor
