+++ todo-app/src/components/ListItem/ListItem.js
@@ -1,17 +1,12 @@
-const ListItem = ({name}) => {
-    // props = {name: 'ivan', surname: 'matas'}
-    // const name = props.name //ivan
-    // const surname = props.surname //ivan
+import Input from "../Shared/Input";
 
+const ListItem = ({ name }) => {
+  return (
+    <div style={{ display: "flex" }}>
+      <Input type="checkbox" />
+      <h3>{name}</h3>
+    </div>
+  );
+};
 
-    // const {name} = props
-    // const name = props.name
-    return (
-        <>
-            <input type="checkbox"/>
-            <h3>{name}</h3>
-        </>
-    )
-}
-
-export default ListItem
+export default ListItem;
