+++ todo-app/src/components/Shared/Input.js
@@ -1,4 +1,4 @@
-const Input = ({ type, value, onChange }) => {
+const Input = ({ type = "text", value, onChange }) => {
   return <input type={type} value={value} onChange={onChange} />;
 };
