+++ todo-app/src/components/TodoList/TodoList.js
@@ -1,11 +1,11 @@
 import { useState } from "react";
-import ListItem from "./ListItem/ListItem";
-import Input from "./Shared/Input";
+import ListItem from "../ListItem/ListItem";
+import Form from "../Shared/Form";
 
 const TodoList = () => {
   const [hasListName, setHasListName] = useState(false);
   const [listName, setListName] = useState("");
-  const [listTodo, setListTodo] = useState([]);
+  const [todoItems, setTodoItems] = useState([]);
   const [todoItem, setTodoItem] = useState("");
 
   const editListNameHandler = () => {
@@ -19,36 +19,32 @@ const TodoList = () => {
 
   const addTodoHandler = (e) => {
     e.preventDefault();
-    setListTodo([...listTodo, todoItem]);
+    setTodoItems([...todoItems, todoItem]);
     setTodoItem("");
   };
 
   return (
     <>
       {!hasListName ? (
-        <form onSubmit={listNameHandler}>
-          <Input
-            type="text"
-            value={listName}
-            onChange={(e) => setListName(e.target.value)}
-          />
-          <button>Create</button>
-        </form>
+        <Form
+          handler={listNameHandler}
+          value={listName}
+          onChange={(e) => setListName(e.target.value)}
+        />
       ) : (
         <div>
           <div>
             <h1>{listName}</h1>
             <button onClick={editListNameHandler}>Edit</button>
           </div>
-          <form onSubmit={addTodoHandler}>
-            <Input
-              value={todoItem}
-              onChange={(e) => setTodoItem(e.target.value)}
-              type="text"
-            />
-            <button>Add</button>
-          </form>
-          {listTodo.map((item, index) => {
+          <br />
+          <h3>Todo Items:</h3>
+          <Form
+            handler={addTodoHandler}
+            value={todoItem}
+            onChange={(e) => setTodoItem(e.target.value)}
+          />
+          {todoItems.map((item, index) => {
             return (
               <div key={index}>
                 <ListItem name={item} />
